refactor ketos.py
split subcommands into separate modules
